// Kulthia Dasboard/src/app/core/models/location.model.ts
export interface LocationModel {
    nonPacket: any;
    usersDetails: any;
    Location: string;
    Series: string;
    LocationCode: string;
    id: any;
    SlNo: any;
    Description: String;
    City: String;
    LocationType: String;
    Address: String;
    ZipCode: String;
    State: String;
}